+++ packages/article-comments/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.1](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.15.0...@times-components/article-comments@0.15.1) (2021-10-13)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
 # [0.15.0](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.14.0...@times-components/article-comments@0.15.0) (2021-10-07)